Improved maximal revision size calculation for ticket #397
git-svn-id: https://svn.code.sf.net/p/numere/source/branches/ticket_397@611 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ common/filerevisions.cpp
@@ -514,9 +514,9 @@ void FileRevisions::fileMove(const wxString& newRevPath, const wxString& comment
 size_t FileRevisions::getMaxDiffFileSize(size_t nFileSize)
 {
     const size_t MINFILESIZE = 512;
-    double dOffset = nFileSize > MINFILESIZE ? MINFILESIZE / (double)nFileSize : 0.5;
-    size_t nTargetFileSize = ((1.0 - dOffset) * 0.8 * exp(-getRevisionCount() / 40.0) + dOffset) * nFileSize;
-    return nTargetFileSize > MINFILESIZE / 2 ? nTargetFileSize : MINFILESIZE / 2;
+    double dOffset = nFileSize > MINFILESIZE ? ((double)MINFILESIZE / (double)nFileSize) : 0.5;
+    size_t nTargetFileSize = ((1.0 - dOffset) * 0.8 * exp(-(double)getRevisionCount() / 40.0) + dOffset) * nFileSize;
+    return nTargetFileSize > (MINFILESIZE / 2) ? nTargetFileSize : (MINFILESIZE / 2);
 }
